logging-materials/local-log-comparison/requirements.txt
pymongo==4.3.3
mysql-connector-python==8.0.32
tabulate==0.9.0
requests==2.31.0
docker==7.0.0
matplotlib==3.8.4